% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{has_bars}
\alias{has_bars}
\alias{find_bars}
\alias{no_bars}
\title{Utilities for formula random effects structure}
\usage{
has_bars(x, type = c("both", "single", "double"))

find_bars(x)

no_bars(x)
}
\arguments{
\item{x}{A formula object}

\item{type}{One of \code{"both"}, \code{"single"}, or \code{"double"}. Defaults to \code{"both"}.}
}
\value{
A modified formula object
}
\description{
Utilities for formula random effects structure
}
\examples{
has_bars(y ~ x)
has_bars(y ~ x + (x | g))
has_bars(y ~ x + (x | g), type = "double")

find_bars(y ~ x)
find_bars(y ~ x + (x | g))

no_bars(y ~ x)
no_bars(y ~ x + (x | g))
}
