% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_formula.R
\name{julia_formula}
\alias{julia_formula}
\title{Convert R formula to Julia syntax}
\usage{
julia_formula(x)
}
\arguments{
\item{x}{A formula object}
}
\value{
A Julia-compatible formula object
}
\description{
Convert R formula to Julia syntax
}
\examples{
julia_formula(y ~ a)

julia_formula(y ~ a + I(a ^ 2) + (a || b))
}
