% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaFormulae-package.R
\docType{package}
\name{JuliaFormulae-package}
\alias{JuliaFormulae}
\alias{JuliaFormulae-package}
\title{JuliaFormulae: Translate R model formulae to 'Julia' syntax}
\description{
Metaprogramming utilites to convert R model formulae to 'Julia' equivalents.
}
\author{
\strong{Maintainer}: June Choe \email{jchoe001@gmail.com} (\href{https://orcid.org/0000-0002-0701-921X}{ORCID})

}
\keyword{internal}
